image/requirements.txt
# API & AWS
boto3==1.34.117
fastapi==0.111.0
mangum==0.17.0
uvicorn==0.30.0

# Document handling & database
pypdf==4.2.0
chromadb==0.5.0

# Environment
python-dotenv==1.0.0

# LangChain - compatible versions
langchain-core==0.1.16
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2.post1

